requirements.txt
# JARVIS AI Assistant Dependencies
# ================================

# =============================================================================
# CORE AI (Online/Offline)
# =============================================================================
# Gemini API (Online Mode)
google-generativeai>=0.8.0

# Memory System (ChromaDB runs on CPU)
chromadb>=0.4.0
sentence-transformers>=2.2.0

# =============================================================================
# PC CONTROL & AUTOMATION
# =============================================================================
pyautogui>=0.9.54
psutil>=5.9.0

# =============================================================================
# VOICE (STT & TTS)
# =============================================================================
# Offline TTS
pyttsx3>=2.90

# Speech Recognition
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

# Wake Word (Offline) - NEW
vosk>=0.3.45

# Enhanced TTS (Optional - works without)
# edge-tts>=6.1.0  # High-quality Microsoft voices (free, no API key)

# =============================================================================
# VISION SYSTEM - NEW
# =============================================================================
# Uses Ollama + LLaVA (installed separately via: ollama pull llava)
Pillow>=10.0.0

# OCR Fallback (Optional - requires Tesseract installed)
# pytesseract>=0.3.10

# =============================================================================
# WEB AUTOMATION - NEW
# =============================================================================
playwright>=1.40.0
# After install, run: playwright install chromium

# =============================================================================
# FILE CONVERSION - NEW
# =============================================================================
PyPDF2>=3.0.0
python-pptx>=0.6.21
python-docx>=1.1.0

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv>=1.0.0
colorama>=0.4.6

# =============================================================================
# CALL SENTRY (Phone Screening) - NEW
# =============================================================================
fastapi>=0.100.0
uvicorn>=0.23.0
twilio>=8.0.0
# websockets>=11.0  # For real-time audio streaming


